JK - Added working playlists to stories
