+++ src/routes/lessons/+page.server.js
@@ -1,11 +1,12 @@
 // playlist/+page.server.js
-import { fetchAllData, mapPlaylistsWithDetails } from '$lib/api';
+import { fetchAllData, mapPlaylistsWithDetails, mapStoriesWithDetails } from '$lib/api';
 import { fetchApi } from '$lib/fetchApi';
 
 export async function load({ locals }) {
     const profileId = 122; //PROFILE ID BECAUSE WE DIDN'T CONNECT EVERYTHING TO DATA YET
     const data = await fetchAllData();
 
+    const storiesWithDetails = mapStoriesWithDetails(data.stories, data.audios, data.languages);
     const enrichedPlaylists = await Promise.all(
         mapPlaylistsWithDetails(data.playlists, data.stories, data.playlistStories).map(async (playlist) => {
             const likes = await fetchApi(`/tm_likes?filter[playlist][_eq]=${playlist.id}&filter[profile][_eq]=${profileId}`);
@@ -20,5 +21,7 @@ export async function load({ locals }) {
     return {
         ...data,
         playlists: enrichedPlaylists,
+        stories: storiesWithDetails,
+        languages: data.languages
     };
 }
